refactor(analyse): move filtering functions
- Moved filtering functions from commands/analyse/analyse.py to
commands/analyse/utils/filtering.py.
- Updated analyse import statements.
